+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -6,7 +6,7 @@ import nextstep.jwp.exception.UncheckedServletException;
 import org.apache.coyote.Processor;
 import org.apache.coyote.http11.message.request.HttpRequest;
 import org.apache.coyote.http11.message.response.HttpResponse;
-import org.apache.coyote.http11.servlet.Servlets;
+import org.apache.catalina.servlets.Servlets;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
